Move response processing to query context
